Msi: 517 refactoring product saleability
