+++ app/code/Magento/InventoryApi/etc/webapi.xml
@@ -69,18 +69,6 @@
             <resource ref="Magento_InventoryApi::stock_delete"/>
         </resources>
     </route>
-    <route url="/V1/inventory/stock/get-product-quantity/:sku/:stockId" method="GET">
-        <service class="Magento\InventorySalesApi\Api\GetProductSalableQtyInterface" method="execute"/>
-        <resources>
-            <resource ref="Magento_InventoryApi::stock"/>
-        </resources>
-    </route>
-    <route url="/V1/inventory/stock/is-product-in-stock/:sku/:stockId" method="GET">
-        <service class="Magento\InventorySalesApi\Api\IsProductSalableInterface" method="execute"/>
-        <resources>
-            <resource ref="Magento_InventoryApi::stock"/>
-        </resources>
-    </route>
     <!-- StockSourceLink -->
     <route url="/V1/inventory/stock/get-assigned-sources/:stockId" method="GET">
         <service class="Magento\InventoryApi\Api\GetAssignedSourcesForStockInterface" method="execute"/>
