+++ app/code/Magento/InventoryCatalog/etc/di.xml
@@ -24,7 +24,7 @@
     </type>
     <type name="Magento\InventoryReservationsApi\Api\AppendReservationsInterface">
         <plugin name="apply_data_to_legacy_catalog_inventory_at_reservation_placing"
-                type="Magento\InventoryCatalog\Plugin\InventoryApi\ApplyDataToLegacyCatalogInventoryAtReservationPlacingPlugin"/>
+                type="Magento\InventoryCatalog\Plugin\InventoryReservationsApi\ApplyDataToLegacyCatalogInventoryAtReservationPlacingPlugin"/>
     </type>
     <type name="Magento\CatalogInventory\Model\ResourceModel\QtyCounterInterface">
         <plugin name="update_source_item_at_legacy_qty_counter" type="Magento\InventoryCatalog\Plugin\CatalogInventory\UpdateSourceItemAtLegacyQtyCounterPlugin"/>
