+++ app/code/Magento/InventorySales/Test/Integration/StockManagement/ReservationPlacingDuringBackItemQtyTest.php
@@ -20,7 +20,7 @@ class ReservationPlacingDuringBackItemQtyTest extends TestCase
     /**
      * @var GetProductSalableQtyInterface
      */
-    private $getProductQtyInStock;
+    private $getProductSalableQty;
 
     /**
      * @var ProductRepositoryInterface
@@ -44,7 +44,7 @@ class ReservationPlacingDuringBackItemQtyTest extends TestCase
 
     protected function setUp()
     {
-        $this->getProductQtyInStock = Bootstrap::getObjectManager()->get(GetProductSalableQtyInterface::class);
+        $this->getProductSalableQty = Bootstrap::getObjectManager()->get(GetProductSalableQtyInterface::class);
         $this->productRepository = Bootstrap::getObjectManager()->get(ProductRepositoryInterface::class);
         $this->stockRepository = Bootstrap::getObjectManager()->get(StockRepositoryInterface::class);
         $this->websiteRepository = Bootstrap::getObjectManager()->get(WebsiteRepositoryInterface::class);
@@ -63,12 +63,12 @@ class ReservationPlacingDuringBackItemQtyTest extends TestCase
      */
     public function testRevertProductsSale()
     {
-        self::assertEquals(8.5, $this->getProductQtyInStock->execute('SKU-1', 10));
+        self::assertEquals(8.5, $this->getProductSalableQty->execute('SKU-1', 10));
 
         $product = $this->productRepository->get('SKU-1');
         $website = $this->websiteRepository->get('eu_website');
         $this->stockManagement->backItemQty($product->getId(), 3.5, $website->getId());
 
-        self::assertEquals(12, $this->getProductQtyInStock->execute('SKU-1', 10));
+        self::assertEquals(12, $this->getProductSalableQty->execute('SKU-1', 10));
     }
 }
