+++ app/code/Magento/InventorySalesApi/etc/webapi.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<routes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Webapi:etc/webapi.xsd">
+    <route url="/V1/inventory/get-product-salable-quantity/:sku/:stockId" method="GET">
+        <service class="Magento\InventorySalesApi\Api\GetProductSalableQtyInterface" method="execute"/>
+        <resources>
+            <resource ref="Magento_InventoryApi::stock"/>
+        </resources>
+    </route>
+    <route url="/V1/inventory/is-product-salable/:sku/:stockId" method="GET">
+        <service class="Magento\InventorySalesApi\Api\IsProductSalableInterface" method="execute"/>
+        <resources>
+            <resource ref="Magento_InventoryApi::stock"/>
+        </resources>
+    </route>
+</routes>
